% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.iMAVEdim.R
\name{cv.iMAVEdim}
\alias{cv.iMAVEdim}
\title{cv-MAVEdim for dimension selection in iMAVE or IMAVE2}
\usage{
cv.iMAVEdim(x, y, tr, dims = c(1:3), pi = 0.5, nfolds = 5, parallel = F, ...)
}
\arguments{
\item{x}{input matrix of dimension nobs x nvars. Each raw is a observation, each column is a covariate}

\item{y}{numeric response}

\item{tr}{is a vector of binary value representing two treatment, 0 or 1.}

\item{dims}{vector of potential dimensions to be reduced}

\item{pi}{is the propensity score p(tr=1|X)}

\item{nfolds}{number of the folds in the cross-validation}

\item{parallel}{implies whether use parallel computing in the cross-validation}

\item{...}{can be other parameters in iMAVE function}
}
\value{
A list
\describe{
\item{cvm}{Mean of the cross-validation for each candidate dimension.}
\item{cvsd}{Estimate of standard error of \code{cvm}.}
\item{cvup}{Upper curve = \code{cvm+cvsd}.}
\item{cvlo}{lower curve = \code{cvm-cvsd}.}
\item{cvraw}{All cross-validation errors on each fold for each candidate dimension.}
\item{iMAVE.fit}{a fitted \code{iMAVE} object with the selected dimension}
\item{beta}{Estimated directions of a fitted \code{iMAVE} object with the selected dimension}
\item{opt_dim}{Selected optimal dimension}
}
#' @references Muxuan Liang, Menggang Yu (2020). A Semiparametric Approach to Model Effect Modification.
}
\description{
cv-MAVEdim for dimension selection in iMAVE or IMAVE2
}
\author{
Muxuan Liang
}
