% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMAVE.R
\name{iMAVE}
\alias{iMAVE}
\title{iMAVE for model effect modification}
\usage{
iMAVE(
  x,
  y,
  tr,
  d = 1,
  pi = 0.5,
  method = "all",
  monoLink = FALSE,
  mainEffect = TRUE,
  initial = "direct",
  tol = 1e-05,
  maxit = 100L,
  normalizeWeight = TRUE,
  constraint = "Grassmann"
)
}
\arguments{
\item{x}{input matrix of dimension nobs x nvars. Each raw is a observation, each column is a covariate}

\item{y}{numeric response}

\item{tr}{is a vector of binary value representing two treatment, 0 or 1.}

\item{d}{number of the dimension to reduce}

\item{pi}{is the propensity score p(tr=1|X)}

\item{method}{'zero' or 'all' decides the local linear expansion only near zero or all data point. 'all' is default for imave and imave2.}

\item{monoLink}{indicates wehter the link function is assumed to be monotone. False by default.}

\item{mainEffect}{mainEffect is TRUE if iMAVE2; FALSE if iMAVE}

\item{initial}{selects the methds to obtain initial values. Default option "direct" start from 0; "zero" solves an imave with method='zero'}

\item{tol}{is the convergence tolerence}

\item{maxit}{is the maxmum iteration}

\item{normalizeWeight}{indicates whether the weight for kernels is normalized. Default is TRUE.}

\item{constraint}{indicates the constaint on B; by default, Grassmann Manifold}
}
\value{
A list
\describe{
\item{dim}{Number of the dimension reduced}
\item{beta}{Estimated linear directions of the dimension reduced on effect mofdification}
\item{ab}{Esimated value and gradient of the link function}
\item{niter}{Number of iterations in optimizaton algorithm to achieve convergence}
\item{mainEffect}{Estimated main effect evaluated at each sample}
}
}
\description{
iMAVE for model effect modification
}
\references{
Muxuan Liang, Menggang Yu (2020). A Semiparametric Approach to Model Effect Modification.
}
\author{
Muxuan Liang
}
