% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iMAVEInfer}
\alias{iMAVEInfer}
\title{A function rather aimed at developers}
\usage{
iMAVEInfer(
  fit,
  x,
  y,
  tr,
  pi,
  mainEffect = NULL,
  sampleSplitIndex = NULL,
  efficient = FALSE
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
