% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMAVEInfer.R
\name{iMAVEInferAugment}
\alias{iMAVEInferAugment}
\title{iMAVEInferAugment for the infetence of the estimated index}
\usage{
iMAVEInferAugment(fit = NULL, x, y, tr, pi = 0.5, efficient = FALSE, ...)
}
\arguments{
\item{fit}{can be a iMAVE object to be used for inference. Defaults run iMAVE internally.}

\item{x}{input matrix of dimension nobs x nvars. Each raw is a observation, each column is a covariate}

\item{y}{numeric response}

\item{tr}{is a vector of binary value representing two treatment, 0 or 1.}

\item{pi}{is the propensity score (success)}

\item{efficient}{imdicates whther use efficient score when the condition variance of the outcome is a constant}

\item{...}{can be other parameters in iMAVE function}
}
\description{
iMAVEInferAugment for the infetence of the estimated index
}
