% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss.score}
\alias{loss.score}
\title{A function rather aimed at developers}
\usage{
loss.score(nfolds, foldid, fitobj, x, y, tr, pi)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
