% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.iMAVEdim.R
\name{cv.iMAVEdim}
\alias{cv.iMAVEdim}
\title{cv-MAVEdim for dimension selection in iMAVE or IMAVE2}
\usage{
cv.iMAVEdim(
  x,
  y,
  tr,
  dims = c(1:ncol(x)),
  pi = 0.5,
  nfolds = 5,
  parallel = F,
  ...
)
}
\arguments{
\item{x}{input matrix of dimension nobs x nvars. Each raw is a observation, each column is a covariate}

\item{y}{numeric response}

\item{tr}{is a vector of binary value representing two treatment, 0 or 1.}

\item{dims}{vector of potential dimensions to be reduced}

\item{pi}{is the propensity score p(tr=1|X)}

\item{nfolds}{number of the folds in the cross-validation}

\item{parallel}{implies whether use parallel computing in the cross-validation}

\item{...}{can be other parameters in iMAVE function}
}
\value{
An object with S3 class "iMAVE"
}
\description{
cv-MAVEdim for dimension selection in iMAVE or IMAVE2
}
